Move invoice methods and tests into merchant
